requirements.txt
# LangGraph RAG Agent 학습 의존성 패키지

# LangGraph 및 LangChain 핵심 패키지
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Vector Store
chromadb>=0.5.0
faiss-cpu>=1.8.0

# 임베딩 및 토크나이저
tiktoken>=0.7.0
sentence-transformers>=3.0.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 유틸리티
pydantic>=2.0.0
httpx>=0.27.0

# 개발 도구 (선택사항)
# pytest>=8.0.0
# black>=24.0.0
